small tweakings

+++ project-nbaudoux-mdevers-womeire/project-nbaudoux-mdevers-womeire/RandomFileCreation.cpp
@@ -2,7 +2,7 @@
 #include "RandomFileCreation.h"
 using namespace std;
 
-RandomFileCreation::RandomFileCreation(string filepath, int length)
+RandomFileCreation::RandomFileCreation(string filepath, size_t length)
 {
 	//Delete file if it existed
 	char filepathChar[_MAX_PATH];
@@ -16,7 +16,7 @@ RandomFileCreation::RandomFileCreation(string filepath, int length)
 	std::default_random_engine generator(rd());
 	std::uniform_int_distribution<int32_t> distribution(std::numeric_limits<int32_t>::min(), std::numeric_limits<int32_t>::max());
 
-	for (int i = 0; i < length; i++)
+	for (size_t i = 0; i < length; i++)
 	{
 		int32_t value = distribution(generator);
 		myfile.write((char *)&value, sizeof(int32_t));
